[CIR] Fixed ClangIR linker
