+++ clang/include/clang/CIR/Dialect/IR/CIRAttrs.td
@@ -972,7 +972,7 @@ def CIR_AddressSpaceAttr :  CIR_EnumAttr<CIR_AddressSpace, "address_space"> {
   ];
 
   let assemblyFormat = [{
-    `` custom<AddressSpaceValue>($value)
+    `<` custom<AddressSpaceValue>($value) `>`
   }];
 
   let defaultValue = "cir::AddressSpace::Default";
